REMOVE: unused Python `scheduling` implementation (#77)
(cherry picked from commit e0703fc7467dc507757b11c99e6b93eb4d848aa3)
